requirements.txt
flask==1.1.2
ShopifyAPI==8.0.2
pyopenssl
